requirements.txt
numpy
opencv-python
matplotlib
pillow
tqdm
ultralytics
